backend/requirements.txt
fastapi>=0.111,<0.116
uvicorn[standard]>=0.29
SQLAlchemy>=2.0
asyncmy>=0.2.9
passlib[bcrypt]>=1.7.4
PyJWT>=2.8
pydantic>=2.7
python-multipart>=0.0.9
aiofiles>=23.2
aiosmtplib>=2.0
tomli>=2.0; python_version < "3.11"
tomli>=2.0
toml>=0.10.2
